Got rid of notebooks, implemented wandb
